Fixes #24060 - fix Puppet environments for host commands
Before this patch, the --environment option in host command was trying
to search in lifecycle environments. Reusing the resolvable definition
from the hostgroup, as they both should have the same behavior.
